Add files via upload
Examine transaction data – look for inconsistencies, missing data across the data set, outliers, correctly identified category items, numeric data across all tables. If any anomaly is detected, make the necessary changes in the dataset and save it. 

Examine customer data – check for similar issues in the customer data, look for nulls and merge the transaction and customer data together so it’s ready for the analysis.

Data analysis and customer segments – look at total sales, drivers of sales, where the highest sales are coming from etc. Explore the data, create charts and graphs as well as noting any interesting trends and/or insights found.

Deep dive into customer segments – derive recommendations from insights, determine which segments should be targeted, if packet sizes are relative and form an overall conclusion based on the analysis. 

Evaluate the performance of a store trial which was performed in stores 77, 86 and 88 - Create a measure to compare different stores to each of the trial stores before the trial period to find control stores with similar performance to each trial store.

Compare each trial and control pair during the trial period. You want to test if total sales are significantly different in the trial period and if so, check if the driver of change is more purchasing customers or more purchases per customers etc.